requirements.txt
certifi==2019.11.28
Django==1.9
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.7.1
pytz==2020.1
wincertstore==0.2
